requirements.txt
requests
pandas
gspread
google-auth
google-auth-oauthlib
python-dotenv
openpyxl
beautifulsoup4
tenacity
